[zen-33] Renamed Consumer to Feature
- Renamed Zend\Module\Consumer to Zend\Module\Feature
- Modified listeners relying on Consumer interfaces:
  - Use Feature interfaces
  - Use combination of typehinting and duck-typing to determine when a
    module is of interest
- Updated and fixed tests to ensure above works
