+++ tests/Zend/Module/TestAsset/ListenerTestModule/Module.php
@@ -1,12 +1,38 @@
 <?php
+/**
+ * Zend Framework
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://framework.zend.com/license/new-bsd
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@zend.com so we can send you a copy immediately.
+ *
+ * @category   Zend
+ * @package    Zend_Module
+ * @subpackage UnitTest
+ * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license    http://framework.zend.com/license/new-bsd     New BSD License
+ */
 
 namespace ListenerTestModule;
 
-use Zend\Module\Consumer\AutoloaderProvider,
-    Zend\Module\Consumer\LocatorRegistered,
-    Zend\Module\Consumer\BootstrapListenerInterface,
+use Zend\Module\Feature\AutoloaderProvider,
+    Zend\Module\Feature\LocatorRegistered,
+    Zend\Module\Feature\BootstrapListenerInterface,
     Zend\EventManager\Event;
 
+/**
+ * @category   Zend
+ * @package    Zend_Module
+ * @subpackage UnitTest
+ * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license    http://framework.zend.com/license/new-bsd     New BSD License
+ */
 class Module implements AutoloaderProvider, LocatorRegistered, BootstrapListenerInterface
 {
     public $initCalled = false;
